Add files via upload
This code when run with Python3 interpreter takes any no of disk image files as input and can carve jpg, png and gif files in the image. This The Carving Path is a 
directory to store retrieved or carved files. 

This tool can be used in Digital Forensics to retrieve images when file table is deleted or corrupted. It can retrieve images that aren't overwritten. 
Any doubts, please reach me on varma0516@gmail.com.
Thank you